Adding max-connection-rate to control connection rate.
Add additional parameter for traffic generation. With this we try to
honor the request rate, and when trying to get a connection to make a
request, we create a new one if the connection rate hasn't been
achieved. If it has been, then we pick up connections from the pool and
if no connections are available in the pool, a new one will be created.

Added a patch for the golang's http transport library to support this.